Threat Analysis Project/ETL_Service/requirements.txt
psycopg2-binary==2.9.9
python-dotenv==1.0.0
geopandas==0.14.1
pandas==1.4.0
sqlalchemy==1.4.25
GeoAlchemy2==0.14.2
numpy==1.22.4
shapely>=1.8.0